PMM-6876 Add Kubernetes cluster status.
