+++ api/managementpb/dbaas/kubernetes.proto
@@ -19,7 +19,7 @@ message KubeAuth {
   ];
 }
 
-// OperatorsStatus defines status of operators installed in kubernetes cluster.
+// OperatorsStatus defines status of operators installed in Kubernetes cluster.
 enum OperatorsStatus {
   // OPERATORS_STATUS_INVALID represents unknown state.
   OPERATORS_STATUS_INVALID = 0;
@@ -27,16 +27,26 @@ enum OperatorsStatus {
   OPERATORS_STATUS_OK = 1;
   // OPERATORS_STATUS_UNSUPPORTED represents that operators are installed, but doesn't have supported API version.
   OPERATORS_STATUS_UNSUPPORTED = 2;
-  // OPERATORS_STATUS_UNAVAILABLE represents that kubernetes cluster is unavailable.
-  OPERATORS_STATUS_UNAVAILABLE = 3;
+  // OPERATORS_STATUS_NOT_INSTALLED represents that operators are not installed.
+  OPERATORS_STATUS_NOT_INSTALLED = 3;
 }
 
-// Operator contains all information about operator installed in kubernetes cluster.
+// KubernetesClusterStatus defines status of Kubernetes cluster.
+enum KubernetesClusterStatus {
+  // KUBERNETES_CLUSTER_STATUS_INVALID represents unknown state.
+  KUBERNETES_CLUSTER_STATUS_INVALID = 0;
+  // KUBERNETES_CLUSTER_STATUS_OK represents that Kubernetes cluster is accessible.
+  KUBERNETES_CLUSTER_STATUS_OK = 1;
+  // KUBERNETES_CLUSTER_STATUS_UNAVAILABLE represents that Kubernetes cluster is not accessible.
+  KUBERNETES_CLUSTER_STATUS_UNAVAILABLE = 2;
+}
+
+// Operator contains all information about operator installed in Kubernetes cluster.
 message Operator {
   OperatorsStatus status = 1;
 }
 
-// Operators contains list of operators installed in kubernetes cluster.
+// Operators contains list of operators installed in Kubernetes cluster.
 message Operators {
   // Percona XtraDB Cluster Operator.
   Operator pxc = 1;
@@ -47,13 +57,15 @@ message Operators {
 message ListKubernetesClustersRequest {}
 
 message ListKubernetesClustersResponse {
-  // Cluster contains public info about kubernetes cluster.
+  // Cluster contains public info about Kubernetes cluster.
   // TODO Do not use inner messages in all public APIs (for consistency).
   message Cluster {
     // Kubernetes cluster name.
     string kubernetes_cluster_name = 1;
     // Installed operators.
     Operators operators = 2;
+    // Status of Kubernetes cluster.
+    KubernetesClusterStatus status = 3;
   }
   // Kubernetes clusters.
   repeated Cluster kubernetes_clusters = 1;
@@ -74,10 +86,7 @@ message RegisterKubernetesClusterRequest {
   ];
 }
 
-message RegisterKubernetesClusterResponse {
-  // Installed operators.
-  Operators operators = 1;
-}
+message RegisterKubernetesClusterResponse {}
 
 message UnregisterKubernetesClusterRequest {
   // Kubernetes cluster name.
